backend/requirements.txt
fastapi
uvicorn[standard]
pandas
weaviate-client==4.5.4
langchain==0.2.7
langchain-core==0.2.15
langchain-community==0.2.7
langchain-weaviate==0.0.5
langchain-huggingface==0.0.3
langchain-openrouter
huggingface-hub==0.23.4
transformers
torch
torchvision
sentence-transformers
python-multipart
scikit-learn
tensorflow
tf-keras
Pillow
requests
notebook
python-dotenv
